Add new Componennt for Token Images (#4496)
