+++ js/src/modals/Transfer/Details/details.js
@@ -14,15 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-import BigNumber from 'bignumber.js';
 import React, { Component, PropTypes } from 'react';
-import { Checkbox, MenuItem } from 'material-ui';
-import { isEqual } from 'lodash';
+import { Checkbox } from 'material-ui';
 
-import Form, { Input, InputAddressSelect, AddressSelect, Select } from '~/ui/Form';
+import Form, { Input, InputAddressSelect, AddressSelect } from '~/ui/Form';
 import { nullableProptype } from '~/util/proptypes';
 
-import imageUnknown from '../../../../assets/images/contracts/unknown-64x64.png';
+import TokenSelect from './tokenSelect';
 import styles from '../transfer.css';
 
 const CHECK_STYLE = {
@@ -31,110 +29,12 @@ const CHECK_STYLE = {
   left: '1em'
 };
 
-class TokenSelect extends Component {
-  static contextTypes = {
-    api: PropTypes.object
-  }
-
-  static propTypes = {
-    onChange: PropTypes.func.isRequired,
-    balance: PropTypes.object.isRequired,
-    images: PropTypes.object.isRequired,
-    tag: PropTypes.string.isRequired
-  };
-
-  componentWillMount () {
-    this.computeTokens();
-  }
-
-  componentWillReceiveProps (nextProps) {
-    const prevTokens = this.props.balance.tokens.map((t) => `${t.token.tag}_${t.value.toNumber()}`);
-    const nextTokens = nextProps.balance.tokens.map((t) => `${t.token.tag}_${t.value.toNumber()}`);
-
-    if (!isEqual(prevTokens, nextTokens)) {
-      this.computeTokens(nextProps);
-    }
-  }
-
-  computeTokens (props = this.props) {
-    const { api } = this.context;
-    const { balance, images } = this.props;
-
-    const items = balance.tokens
-      .filter((token, index) => !index || token.value.gt(0))
-      .map((balance, index) => {
-        const token = balance.token;
-        const isEth = index === 0;
-        let imagesrc = token.image;
-
-        if (!imagesrc) {
-          imagesrc =
-            images[token.address]
-              ? `${api.dappsUrl}${images[token.address]}`
-              : imageUnknown;
-        }
-        let value = 0;
-
-        if (isEth) {
-          value = api.util.fromWei(balance.value).toFormat(3);
-        } else {
-          const format = balance.token.format || 1;
-          const decimals = format === 1 ? 0 : Math.min(3, Math.floor(format / 10));
-
-          value = new BigNumber(balance.value).div(format).toFormat(decimals);
-        }
-
-        const label = (
-          <div className={ styles.token }>
-            <img src={ imagesrc } />
-            <div className={ styles.tokenname }>
-              { token.name }
-            </div>
-            <div className={ styles.tokenbalance }>
-              { value }<small> { token.tag }</small>
-            </div>
-          </div>
-        );
-
-        return (
-          <MenuItem
-            key={ `${index}_${token.tag}` }
-            value={ token.tag }
-            label={ label }
-          >
-            { label }
-          </MenuItem>
-        );
-      });
-
-    this.setState({ items });
-  }
-
-  render () {
-    const { tag, onChange } = this.props;
-    const { items } = this.state;
-
-    return (
-      <Select
-        className={ styles.tokenSelect }
-        label='type of token transfer'
-        hint='type of token to transfer'
-        value={ tag }
-        onChange={ onChange }
-      >
-        { items }
-      </Select>
-    );
-  }
-}
-
 export default class Details extends Component {
   static propTypes = {
     address: PropTypes.string,
     balance: PropTypes.object,
     all: PropTypes.bool,
     extras: PropTypes.bool,
-    images: PropTypes.object.isRequired,
     sender: PropTypes.string,
     senderError: PropTypes.string,
     sendersBalances: PropTypes.object,
@@ -249,12 +149,11 @@ export default class Details extends Component {
   }
 
   renderTokenSelect () {
-    const { balance, images, tag } = this.props;
+    const { balance, tag } = this.props;
 
     return (
       <TokenSelect
         balance={ balance }
-        images={ images }
         tag={ tag }
         onChange={ this.onChangeToken }
       />
