+++ js/src/modals/Transfer/transfer.js
@@ -44,7 +44,6 @@ class Transfer extends Component {
   static propTypes = {
     newError: PropTypes.func.isRequired,
     gasLimit: PropTypes.object.isRequired,
-    images: PropTypes.object.isRequired,
 
     senders: nullableProptype(PropTypes.object),
     sendersBalances: nullableProptype(PropTypes.object),
@@ -174,7 +173,7 @@ class Transfer extends Component {
   }
 
   renderDetailsPage () {
-    const { account, balance, images, senders } = this.props;
+    const { account, balance, senders } = this.props;
     const { recipient, recipientError, sender, senderError, sendersBalances } = this.store;
     const { valueAll, extras, tag, total, totalError, value, valueError } = this.store;
 
@@ -184,7 +183,6 @@ class Transfer extends Component {
         all={ valueAll }
         balance={ balance }
         extras={ extras }
-        images={ images }
         onChange={ this.store.onUpdateDetails }
         recipient={ recipient }
         recipientError={ recipientError }
