+++ js/src/views/Account/account.js
@@ -37,7 +37,6 @@ class Account extends Component {
   static propTypes = {
     fetchCertifiers: PropTypes.func.isRequired,
     fetchCertifications: PropTypes.func.isRequired,
-    images: PropTypes.object.isRequired,
     setVisibleAccounts: PropTypes.func.isRequired,
 
     accounts: PropTypes.object,
@@ -257,14 +256,13 @@ class Account extends Component {
       return null;
     }
 
-    const { balances, images } = this.props;
+    const { balances } = this.props;
 
     return (
       <Transfer
         account={ account }
         balance={ balance }
         balances={ balances }
-        images={ images }
         onClose={ this.store.toggleTransferDialog }
       />
     );
@@ -289,12 +287,10 @@ class Account extends Component {
 function mapStateToProps (state) {
   const { accounts } = state.personal;
   const { balances } = state.balances;
-  const { images } = state;
 
   return {
     accounts,
-    balances,
-    images
+    balances
   };
 }
